expenses crud

+++ database/migrations/2021_12_09_1639046855_create_debts_table.php
@@ -17,7 +17,7 @@ class CreateDebtsTable extends Migration
         Schema::create('debts', function (Blueprint $table) {
             $table->bigIncrements('id');
 $table->foreignId("admin_id")->constrained("admins")->onUpdate("cascade")->onDelete("cascade");
-            $table->bigInteger('amount');
+            $table->double('amount');
             $table->string('note')->nullable();
             $table->foreignId("employee_id")->constrained("employees")->references("id");
 			$table->softDeletes();
